cleaning and changing structure
