+++ lua/config/autocmds.lua
@@ -134,10 +134,7 @@ vim.api.nvim_create_autocmd("BufEnter", {
 
 vim.api.nvim_create_autocmd({ "BufWritePost" }, {
   callback = function()
-    -- try_lint without arguments runs the linters defined in linters_by_ft
-    -- for the current filetype
     require("lint").try_lint()
-    -- require("lint").try_lint("gfortran")
   end,
 })
