+++ lua/plugins/ai.lua
@@ -17,6 +17,7 @@ return {
   },
   {
     "CopilotC-Nvim/CopilotChat.nvim",
+    enabled = false,
     branch = "main",
     cmd = "CopilotChat",
     opts = function()
@@ -78,6 +79,7 @@ return {
   },
   {
     "zbirenbaum/copilot.lua",
+    enabled = false,
     cmd = "Copilot",
     event = "InsertEnter",
     config = function()
