+++ lua/plugins/management.lua
@@ -1,0 +1,60 @@
+return {
+  { -- project auto management
+    "ahmedkhalf/project.nvim",
+    event = "VeryLazy",
+    opts = {
+      active = true,
+      on_config_done = nil,
+      manual_mode = false,
+      detection_methods = { "pattern", "lsp" },
+      patterns = { ".git" },
+      show_hidden = false,
+      silent_chdir = true,
+      ignore_lsp = {},
+      sync_root_with_cwd = true,
+      respect_buf_cwd = true,
+      update_focused_file = {
+        enable = true,
+        update_root = true,
+      },
+    },
+    config = function(opts)
+      require("project_nvim").setup(opts)
+    end,
+  },
+  { -- session auto management
+    "folke/persistence.nvim",
+    event = "BufReadPre",
+    opts = {},
+    keys = {
+      {
+        "<leader>qs",
+        function()
+          require("persistence").load()
+        end,
+        desc = "Restore Session",
+      },
+      {
+        "<leader>qS",
+        function()
+          require("persistence").select()
+        end,
+        desc = "Select Session",
+      },
+      {
+        "<leader>ql",
+        function()
+          require("persistence").load({ last = true })
+        end,
+        desc = "Restore Last Session",
+      },
+      {
+        "<leader>qd",
+        function()
+          require("persistence").stop()
+        end,
+        desc = "Don't Save Current Session",
+      },
+    },
+  },
+}
