+++ lua/plugins/mini.lua
@@ -1,0 +1,28 @@
+return {
+  {
+    "echasnovski/mini.nvim",
+    event = { "BufReadPre", "BufNewFile" },
+    config = function()
+      require("mini.ai").setup({ n_lines = 500 })
+      require("mini.surround").setup({
+        mappings = {
+          add = "gsa",
+          delete = "gsd",
+          find = "gsf",
+          find_left = "gsF",
+          highlight = "gsh",
+          replace = "gsr",
+          update_n_lines = "gsn",
+        },
+      })
+      require("mini.icons").setup()
+      require("mini.comment").setup({
+        options = {
+          custom_commentstring = function()
+            return require("ts_context_commentstring.internal").calculate_commentstring() or vim.bo.commentstring
+          end,
+        },
+      })
+    end,
+  },
+}
