+++ lua/plugins/treesitter.lua
@@ -1,0 +1,97 @@
+return {
+  { -- incrementally select a part of code with enter
+    "sustech-data/wildfire.nvim",
+    event = { "BufReadPre", "BufNewFile" },
+    dependencies = { "nvim-treesitter/nvim-treesitter" },
+    config = function()
+      require("wildfire").setup()
+    end,
+  },
+  {
+    "nvim-treesitter/nvim-treesitter",
+    event = { "BufReadPre", "BufNewFile" },
+    build = ":TSUpdate",
+    cmd = { "TSInstall", "TSUpdate", "TSUpdateSync" },
+    dependencies = {
+      "nvim-treesitter/nvim-treesitter-textobjects",
+    },
+    opts = {
+      ensure_installed = {
+        "bash",
+        "html",
+        "json",
+        "lua",
+        "markdown",
+        "markdown_inline",
+        "python",
+        "vim",
+        "vimdoc",
+        "yaml",
+        "toml",
+        "julia",
+        "fortran",
+        "fish",
+        "csv",
+        "diff",
+      },
+      auto_install = true,
+      indent = {
+        enable = true,
+        disable = { "julia", "fortran" },
+      },
+      highlight = {
+        enable = true,
+      },
+      textobjects = {
+        move = {
+          enable = true,
+          goto_next_start = {
+            ["]f"] = "@function.outer",
+            ["]c"] = "@class.outer",
+            ["]a"] = "@parameter.inner",
+            ["]b"] = "@block.outer",
+          },
+          goto_next_end = {
+            ["]F"] = "@function.outer",
+            ["]C"] = "@class.outer",
+            ["]A"] = "@parameter.inner",
+            ["]B"] = "@block.outer",
+          },
+          goto_previous_start = {
+            ["[f"] = "@function.outer",
+            ["[c"] = "@class.outer",
+            ["[a"] = "@parameter.inner",
+            ["[b"] = "@block.outer",
+          },
+          goto_previous_end = {
+            ["[F"] = "@function.outer",
+            ["[C"] = "@class.outer",
+            ["[A"] = "@parameter.inner",
+            ["[B"] = "@block.outer",
+          },
+        },
+        select = {
+          enable = true,
+          keymaps = {
+            ["af"] = "@function.outer",
+            ["if"] = "@function.inner",
+            ["aC"] = "@class.outer",
+            ["iC"] = "@class.inner",
+            ["al"] = "@loop.outer",
+            ["il"] = "@loop.outer",
+            ["ac"] = "@conditional.outer",
+            ["ic"] = "@conditional.inner",
+            ["ao"] = "@block.outer",
+            ["io"] = "@block.inner",
+            -- ["ic"] = { query = "@class.inner", desc = "Select inner part of a class region" },
+          },
+        },
+      },
+    },
+    config = function(_, opts)
+      require("nvim-treesitter.install").prefer_git = true
+      ---@diagnostic disable-next-line: missing-fields
+      require("nvim-treesitter.configs").setup(opts)
+    end,
+  },
+}
