+++ lua/plugins/ui.lua
@@ -1,0 +1,60 @@
+return {
+  { -- bunch of useful ui
+    "folke/snacks.nvim",
+    lazy = false,
+    opts = {
+      dashboard = {
+        preset = {
+          header = [[
+                                                          ⠀⠀⠀⠀⠀⠀⣀⡤⠴⠒⠒⠒⠶⢤⣄⡀⠀⠀⠀⠀⠀⠀           
+██████╗ ███████╗███╗   ██╗ ██████╗ ██╗   ██╗██╗███╗   ███╗⠀⠀⠀⠀⣠⠞⠁⠀⠀⠀⠀⠀⠀⠀⠈⠙⢦⡀⠀⠀⠀⠀           
+██╔══██╗██╔════╝████╗  ██║██╔════╝ ██║   ██║██║████╗ ████║⠀⠀⠀⡼⠁⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⢳⡀⠀< Hello World
+██████╔╝█████╗  ██╔██╗ ██║██║  ███╗██║   ██║██║██╔████╔██║⠀⠀⢸⠁⠀⠀⣠⠖⠛⠛⠲⢤⠀⠀⠀⣰⠚⠛⢷⠀⠀⠀           
+██╔═══╝ ██╔══╝  ██║╚██╗██║██║   ██║╚██╗ ██╔╝██║██║╚██╔╝██║⠀⠀⣿⠀⠀⣸⠃⠀⠀⢀⣀⠈⢧⣠⣤⣯⢠⣤⠘⣆⠀⠀           
+██║     ███████╗██║ ╚████║╚██████╔╝ ╚████╔╝ ██║██║ ╚═╝ ██║⠀⠀⣿⠀⠀⡇⠀⠀⠀⠻⠟⠠⣏⣀⣀⣨⡇⠉⢀⣿⠀⠀           
+╚═╝     ╚══════╝╚═╝  ╚═══╝ ╚═════╝   ╚═══╝  ╚═╝╚═╝     ╚═╝⠀⢀⡟⠀⠀⠹⡄⠀⠀⠀⠀⠀⠉⠑⠚⠉⠀⣠⡞⢿⠀⠀           
+                                                          ⢀⡼⠁⠀⠀⠀⠙⠳⢤⡄⠀⠀⠀⠀⠀⠀⠀⠁⠙⢦⠳⣄           
+    ]],
+          keys = {
+            { action = ":FzfLua files", desc = " Find File", icon = " ", key = "f" },
+            { action = ":ene | startinsert", desc = " New File", icon = " ", key = "n" },
+            { action = ":FzfLua oldfiles", desc = " Recent Files", icon = " ", key = "r" },
+            { action = ":FzfLua live_grep", desc = " Find Text", icon = " ", key = "g" },
+            {
+              action = ':lua require("persistence").load()',
+              desc = " Restore Session",
+              icon = " ",
+              key = "s",
+            },
+            { action = ":Lazy", desc = " Lazy", icon = "󰒲 ", key = "l" },
+            { action = ":qa", desc = " Quit", icon = " ", key = "q" },
+          },
+        },
+      },
+      statuscolumn = {
+        enabled = false,
+      },
+    },
+  },
+  {
+    "norcalli/nvim-colorizer.lua",
+    lazy = true,
+    config = function()
+      require("colorizer").setup({})
+    end,
+    keys = {
+      {
+        "<leader>uH",
+        "<cmd>ColorizerToggle<CR>",
+        desc = "Colorizer Toggle",
+      },
+    },
+  },
+  {
+    "OXY2DEV/helpview.nvim",
+    ft = "help",
+    dependencies = {
+      "nvim-treesitter/nvim-treesitter",
+    },
+  },
+}
