+++ lua/plugins/whichkey.lua
@@ -1,0 +1,100 @@
+return {
+  {
+    "folke/which-key.nvim",
+    opts_extend = { "spec" },
+    opts = {
+      ---@type wk.Win
+      win = {
+        height = { min = 4, max = 25 },
+        col = 0.5,
+        row = -1,
+        border = "double",
+        padding = { 0, 1 },
+        title = false,
+        title_pos = "center",
+        zindex = 1000,
+        bo = {},
+        wo = {},
+      },
+      layout = {
+        width = { min = 20 },
+        spacing = 1,
+        align = "left",
+      },
+      show_keys = false,
+      show_help = false,
+      preset = "classic",
+      defaults = {},
+      spec = {
+        {
+          mode = { "n", "v" },
+          { "<leader><tab>", group = "tabs" },
+          { "<leader>c", group = "code" },
+          { "<leader>d", group = "debug" },
+          { "<leader>dp", group = "profiler" },
+          { "<leader>f", group = "file/find" },
+          { "<leader>g", group = "git" },
+          { "<leader>gd", group = "diffview" },
+          { "<leader>gh", group = "hunks" },
+          { "<leader>q", group = "quit/session" },
+          { "<leader>s", group = "search" },
+          { "<leader>u", group = "ui", icon = { icon = "󰙵 ", color = "cyan" } },
+          { "<leader>x", group = "diagnostics/quickfix", icon = { icon = "󱖫 ", color = "green" } },
+          { "[", group = "prev" },
+          { "]", group = "next" },
+          { "g", group = "goto" },
+          { "gs", group = "surround" },
+          { "z", group = "fold" },
+          {
+            "<leader>b",
+            group = "buffer",
+            expand = function()
+              return require("which-key.extras").expand.buf()
+            end,
+          },
+          {
+            "<leader>w",
+            group = "windows",
+            proxy = "<c-w>",
+            expand = function()
+              return require("which-key.extras").expand.win()
+            end,
+          },
+          { "gx", desc = "Open with system app" },
+        },
+      },
+    },
+    keys = {
+      {
+        "<leader>?",
+        function()
+          require("which-key").show({ global = false })
+        end,
+        desc = "Buffer Keymaps (which-key)",
+      },
+      {
+        "<c-w><space>",
+        function()
+          require("which-key").show({ keys = "<c-w>", loop = true })
+        end,
+        desc = "Window Hydra Mode (which-key)",
+      },
+    },
+    config = function(_, opts)
+      local wk = require("which-key")
+      wk.setup(opts)
+      wk.add({
+        {
+          { "<leader>w", group = "window" },
+          { "<leader>wh", "<C-W>h", desc = "Move to left" },
+          { "<leader>wj", "<C-W>j", desc = "Move down" },
+          { "<leader>wk", "<C-W>k", desc = "Move up" },
+          { "<leader>wl", "<C-W>l", desc = "Move to right" },
+          { "<leader>z", group = "Zoom" },
+          { "<leader>zi", "<cmd>tab split<CR>", desc = "Zoom in" },
+          { "<leader>zo", "<cmd>tab close<CR>", desc = "Zoom out" },
+        },
+      })
+    end,
+  },
+}
